methods added to messages

+++ messenger/src/main/java/org/acc/messenger/model/Message.java
@@ -9,7 +9,7 @@ import javax.xml.bind.annotation.XmlRootElement;
 public class Message {
 
 	
-	private int id;
+	private long id;
 	private String message;
 	private Date created;
 	private String author;
@@ -18,17 +18,17 @@ public class Message {
 		
 		
 	}
-	public Message(int id, String message, String author) {
+	public Message(long id, String message, String author) {
 		super();
 		this.id = id;
 		this.message = message;
 		this.author = author;
 		this.created = new Date();
 	}
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getMessage() {
